Refactoring to IEnumerable result and unit tests
