+++ CDP4Common/Poco/Option.cs
@@ -24,6 +24,8 @@
 
 namespace CDP4Common.EngineeringModelData
 {
+    using System;
+    using System.Collections.Generic;
     using System.Linq;
 
     using CDP4Common.Helpers;
@@ -47,20 +49,20 @@ namespace CDP4Common.EngineeringModelData
         }
 
         /// <summary>
-        /// Finds a NestedParameter by its <see cref="NestedParameter.Path"/> and returns its <see cref="NestedParameter.ActualValue"/>
-        /// as the generic <typeparamref name="T"></typeparamref>'s .
+        /// Finds a NestedParameter by its <see cref="NestedParameter.Path"/> in this <see cref="Option"/>'s <see cref="NestedParameter"/>s
+        /// and returns its <see cref="NestedParameter.ActualValue"/> "converted" to the generic <typeparamref name="T"></typeparamref>'s .
         /// </summary>
-        /// <typeparam name="T">The generic type to which the <see cref="NestedParameter.ActualValue"/> needs to be converted.</typeparam>
-        /// <param name="path">The path to search for in all <see cref="NestedParameter.Path"/> properties.</param>
+        /// <typeparam name="T">The generic type to which the <see cref="NestedParameter.ActualValue"/> needs to be "converted".</typeparam>
+        /// <param name="path">The path to search for in all this <see cref="Option"/>'s <see cref="NestedParameter.Path"/> properties.</param>
         /// <param name="domain">The <see cref="DomainOfExpertise"/> for which the <see cref="NestedParameter"/>s should nbe found.</param>
         /// <returns>A single <see cref="NestedParameter"/> if the path was found
         /// and its <see cref="NestedParameter.ActualValue"/> could be converted to the requested generic <typeparamref name="T"></typeparamref>, otherwise null.</returns>
-        public T GetNestedParameterValueByPath<T>(string path, DomainOfExpertise domain) where T : class
+        public IEnumerable<T> GetNestedParameterValuesByPath<T>(string path, DomainOfExpertise domain)
         {
             var allParameters = new NestedElementTreeGenerator().GetNestedParameters(this, domain);
-            var nestedParameter = allParameters.SingleOrDefault(x => x.Path.Equals(path));
 
-            return nestedParameter?.ActualValue.ToValueSetObject(nestedParameter.AssociatedParameter.ParameterType) as T;
+            return allParameters.Where(x => x.Path.Equals(path))
+                .Select(x => (T)Convert.ChangeType(x.ActualValue.ToValueSetObject(x.AssociatedParameter.ParameterType), typeof(T))).ToArray();
         }
     }
 }
